GameDanhBai/requirements.txt
fastapi
uvicorn
python-socketio
jinja2
python-multipart
aiofiles
websockets